+++ src/services/dashboards/widgets/_configs/continent-config.ts
