+++ src/services/dashboards/widgets/cost-by-region/CostByRegion.vue
@@ -69,12 +69,12 @@ import WidgetDataTable from '@/services/dashboards/widgets/_components/WidgetDat
 import WidgetFrame from '@/services/dashboards/widgets/_components/WidgetFrame.vue';
 import type { WidgetExpose, WidgetProps } from '@/services/dashboards/widgets/_configs/config';
 import { GROUP_BY } from '@/services/dashboards/widgets/_configs/config';
+import { CONTINENT_INFO } from '@/services/dashboards/widgets/_configs/continent-config';
 import { getLegends } from '@/services/dashboards/widgets/_helpers/widget-chart-helper';
 import { useWidgetFrameProps } from '@/services/dashboards/widgets/_hooks/use-widget-frame-props';
 import { useWidgetLifecycle } from '@/services/dashboards/widgets/_hooks/use-widget-lifecycle';
 // eslint-disable-next-line import/no-cycle
 import { useWidgetState } from '@/services/dashboards/widgets/_hooks/use-widget-state';
-import { CONTINENT_INFO } from '@/services/dashboards/widgets/cost-by-region/config';
 import type { AccumulatedDataModel, Legend } from '@/services/dashboards/widgets/type';
 
 type Data = AccumulatedDataModel['results'];
